betfair-arbitrage-tool/requirements.txt
betfairlightweight==2.19.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
beautifulsoup4==4.12.2
lxml==4.9.3
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
loguru==0.7.2
pydantic==2.5.2
aiohttp==3.9.1
asyncio==3.4.3
apscheduler==3.10.4
eventlet==0.33.3